ticket430 add fix to update on observer

+++ blocks/import_remote_course/classes/subscriber.php
@@ -87,7 +87,9 @@ class subscriber {
                 $DB->insert_record($table, $dataobject);
                 break;
             case 'd':
-                $DB->delete_records($table, array('course_id' => $course_id));
+            	if ($DB->get_field($table, 'id', array('course_id' => $course_id,))) {
+                	$DB->delete_records($table, array('course_id' => $course_id));
+            	}
                 break;
             case 'u':
                 $id = $DB->get_field($table, 'id', array('course_id' => $course_id,));
